doc: Add some docs.
