+++ crates/bevy-inspector-egui/src/minibuffer/filter_query_inspector.rs
@@ -1,3 +1,22 @@
+//! # filter_query_inspector act
+//!
+//! ## Usage
+//!
+//! ```no_run
+//! use bevy::prelude::*;
+//! use bevy_minibuffer::prelude::*;
+//! use bevy_inspector_egui::minibuffer;
+//! fn plugin(app: &mut App) {
+//!     app
+//!         .add_plugins(MinibufferPlugins)
+//!         .add_acts((
+//!             BasicActs::default(),
+//!             minibuffer::FilterQueryInspectorActs::default()
+//!                 .add::<With<Transform>>()
+//!                 .add::<With<Mesh3d>>()
+//!         ));
+//! }
+//! ```
 use crate::{
     minibuffer::{InspectorPlugins, Inspectors},
     quick::FilterQueryInspectorPlugin,
@@ -12,6 +31,8 @@ use bevy_ecs::{
 use bevy_minibuffer::{prelude::*, prompt::PromptState};
 use bevy_state::prelude::in_state;
 
+/// Adds the 'filter_query_inspector' act which toggles the visibility of the
+/// added filter query inspectors.
 pub struct FilterQueryInspectorActs {
     plugins: InspectorPlugins<Self>,
     acts: Acts,
